refactor: reorder logging
