+++ packages/repack/src/logging/index.ts
@@ -1,5 +1,3 @@
-export * from './reporters/FileReporter.js';
-export * from './reporters/ConsoleReporter.js';
-export * from './compose.js';
+export * from './helpers.js';
+export * from './reporters.js';
 export * from './types.js';
-export * from './makeLogEntryFromFastifyLog.js';
